+++ packages/dns-discovery/src/dns_over_https.ts
@@ -1,4 +1,4 @@
-import { bytesToUtf8 } from "@waku/byte-utils";
+import { bytesToUtf8 } from "@waku/utils";
 import debug from "debug";
 import { Endpoint, query, toEndpoint } from "dns-query";
 
@@ -16,10 +16,10 @@ export class DnsOverHttps implements DnsClient {
    */
   static DefaultEndpoints: Endpoint[] = [
     toEndpoint({
-      name: "cisco-doh",
+      name: "AhaDNS",
       protocol: "https:",
-      host: "doh.opendns.com",
-      ipv4: "146.112.41.2",
+      host: "doh.la.ahadns.net",
+      ipv4: "45.67.219.208",
     }),
     toEndpoint({
       name: "cloudflare",