+++ packages/message-encryption/src/crypto/index.ts
@@ -1,7 +1,7 @@
 import nodeCrypto from "crypto";
 
 import * as secp from "@noble/secp256k1";
-import { concat } from "@waku/byte-utils";
+import { concat } from "@waku/utils";
 import sha3 from "js-sha3";
 
 import { Asymmetric, Symmetric } from "../constants.js";