+++ packages/tests/tests/peer_exchange.node.spec.ts
@@ -16,6 +16,15 @@ import { Nwaku } from "../src/nwaku.js";
 
 describe("Peer Exchange", () => {
   let waku: LightNode;
+
+  before(async function () {
+    // skipping in CI as this test demonstrates Peer Exchange working with the test fleet
+    // but not with locally run nwaku nodes
+    if (process.env.CI) {
+      this.skip();
+    }
+  });
+
   afterEach(async function () {
     !!waku && waku.stop().catch((e) => console.log("Waku failed to stop", e));
   });
@@ -33,7 +42,7 @@ describe("Peer Exchange", () => {
     });
 
     await waku.start();
-    await delay(1000);
+    await delay(100000);
 
     await waitForRemotePeer(waku, [Protocols.PeerExchange]);
     const pxPeers = await waku.peerExchange.peers();
@@ -42,12 +51,6 @@ describe("Peer Exchange", () => {
 
   it("Manual query on test fleet", async function () {
     this.timeout(150_000);
-
-    // skipping in CI as this test demonstrates Peer Exchange working with the test fleet
-    // but not with locally run nwaku nodes
-    if (process.env.ci) {
-      this.skip();
-    }
 
     const waku = await createLightNode({
       libp2p: {